# Remove is_diagonal, is_(lower|upper)_triangular for ZZMatrix