The generic methods use `is_zero_entry` and thus have a
performance profile similar to the functions being removed here.

Before:

    julia> A = diagonal_matrix(ZZ(42), 10);

    julia> @btime is_diagonal(A);
      208.696 ns (0 allocations: 0 bytes)

    julia> @btime is_lower_triangular(A);
      125.095 ns (0 allocations: 0 bytes)

    julia> @btime is_upper_triangular(A);
      127.596 ns (0 allocations: 0 bytes)

    julia> A = diagonal_matrix(ZZ(42), 100);

    julia> @btime is_diagonal(A);
      19.750 μs (0 allocations: 0 bytes)

    julia> @btime is_lower_triangular(A);
      11.083 μs (0 allocations: 0 bytes)

    julia> @btime is_upper_triangular(A);
      11.125 μs (0 allocations: 0 bytes)

After:

    julia> A = diagonal_matrix(ZZ(42), 10);

    julia> @btime is_diagonal(A);
      208.708 ns (0 allocations: 0 bytes)

    julia> @btime is_lower_triangular(A);
      129.279 ns (0 allocations: 0 bytes)

    julia> @btime is_upper_triangular(A);
      124.396 ns (0 allocations: 0 bytes)

    julia> A = diagonal_matrix(ZZ(42), 100);

    julia> @btime is_diagonal(A);
      19.375 μs (0 allocations: 0 bytes)

    julia> @btime is_lower_triangular(A);
      11.166 μs (0 allocations: 0 bytes)

    julia> @btime is_upper_triangular(A);
      11.125 μs (0 allocations: 0 bytes)